API/app.py
from flask import Flask, request
from google.oauth2.credentials import Credentials
import boto3
from azure.identity import ClientSecretCredential
import json

app = Flask(__name__)

# Dictionary to store user credentials and code
user_data = {}

@app.route('/authenticate', methods=['POST'])
def authenticate():
    # Get user identifier, provider, and credentials from request
    data = request.get_json()
    user_id = data['user_id']
    provider = data['provider']
    credentials = data['credentials']

    # Store user provider and credentials in dictionary
    if user_id not in user_data:
        user_data[user_id] = {}
    user_data[user_id]['provider'] = provider
    user_data[user_id]['credentials'] = credentials

    return 'Authenticated'

@app.route('/run', methods=['POST'])
def run():
    # Get user identifier and code from request
    data = request.get_json()
    user_id = data['user_id']
    code = data['code']

    # Store user code in dictionary
    if user_id not in user_data:
        user_data[user_id] = {}
    user_data[user_id]['code'] = code

    # Get user provider and credentials from dictionary
    provider = user_data[user_id]['provider']
    credentials = user_data[user_id]['credentials']

    # Authenticate with cloud provider
    if provider == 'gcp':
        # Authenticate with GCP using OAuth 2.0 credentials
        creds = Credentials.from_authorized_user_info(info=credentials)
        # ...
    elif provider == 'aws':
        # Authenticate with AWS using access key and secret access key
        session = boto3.Session(
            aws_access_key_id=credentials['aws_access_key_id'],
            aws_secret_access_key=credentials['aws_secret_access_key'],
        )
        # ...
    elif provider == 'azure':
        # Authenticate with Azure using client ID, client secret, and tenant ID
        creds = ClientSecretCredential(
            tenant_id=credentials['tenant_id'],
            client_id=credentials['client_id'],
            client_secret=credentials['client_secret'],
        )
        # ...

    # Run CDKTF code
    exec(code)

    return 'Code executed'

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=8000)