API/requirements.txt
Flask==2.1.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
boto3==1.20.24
azure-identity==1.7.1
cdktf==0.8.2